refactor export entries filename
